Refactor and abstract PreviewSurfacce control
