+++ Neo/Parcel.Neo/Parcel.Neo.csproj.user
@@ -8,6 +8,9 @@
     <Compile Update="PopupWindows\ActionEventPickerWindow.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Update="PreviewSurface.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Update="Prompts\StringEntryPromptDialog.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -26,6 +29,9 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Update="PopupWindows\CreateFunctionWindow.xaml">
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Update="PreviewSurface.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Update="Prompts\StringEntryPromptDialog.xaml">
